+++ src/main/java/net/chaosgames/ringchaos/datagen/ModCuriosProvider.java
@@ -1,0 +1,20 @@
+package net.chaosgames.ringchaos.datagen;
+
+import net.minecraft.core.HolderLookup;
+import net.minecraft.data.PackOutput;
+import net.minecraftforge.common.data.ExistingFileHelper;
+import top.theillusivec4.curios.api.CuriosDataProvider;
+
+import java.util.concurrent.CompletableFuture;
+
+public class ModCuriosProvider extends CuriosDataProvider {
+    public ModCuriosProvider(String modId, PackOutput output, ExistingFileHelper fileHelper,
+                             CompletableFuture<HolderLookup.Provider> registries) {
+        super(modId, output, fileHelper, registries);
+    }
+
+    @Override
+    public void generate(HolderLookup.Provider registries, ExistingFileHelper fileHelper) {
+        this.createEntities("ringchaos_entities").addPlayer().addSlots("ring");
+    }
+}